Move loadability/resolvability checks for methods to ContainsMethodBody
