+++ src/coreclr/tools/aot/ILCompiler.ReadyToRun/Compiler/ReadyToRunLibraryRootProvider.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Collections.Generic;
+using System.Reflection.Metadata;
 
 using Internal.TypeSystem.Ecma;
 using Internal.TypeSystem;
@@ -63,7 +64,7 @@ namespace ILCompiler
                 {
                     if (!CorInfoImpl.ShouldSkipCompilation(_instructionSetSupport, method))
                     {
-                        DependencyAnalysis.NodeFactory.CheckCanGenerateMethod(methodToRoot);
+                        CheckCanGenerateMethod(methodToRoot);
                         rootProvider.AddCompilationRoot(methodToRoot, rootMinimalDependencies: false, reason: reason);
                     }
                 }
@@ -74,6 +75,52 @@ namespace ILCompiler
                     continue;
                 }
             }
+        }
+
+        /// <summary>
+        /// Validates that it will be possible to generate '<paramref name="method"/>' based on the types
+        /// in its signature. Unresolvable types in a method's signature prevent RyuJIT from generating
+        /// even a stubbed out throwing implementation.
+        /// </summary>
+        public static void CheckCanGenerateMethod(MethodDesc method)
+        {
+            // Ensure the method is loadable
+            ((CompilerTypeSystemContext)method.Context).EnsureLoadableMethod(method);
+
+            MethodSignature signature = method.Signature;
+
+            // Vararg methods are not supported in .NET Core
+            if ((signature.Flags & MethodSignatureFlags.UnmanagedCallingConventionMask) == MethodSignatureFlags.CallingConventionVarargs)
+                ThrowHelper.ThrowBadImageFormatException();
+
+            CheckTypeCanBeUsedInSignature(signature.ReturnType);
+
+            for (int i = 0; i < signature.Length; i++)
+            {
+                CheckTypeCanBeUsedInSignature(signature[i]);
+            }
+        }
+
+        private static void CheckTypeCanBeUsedInSignature(TypeDesc type)
+        {
+            DefType defType = type as DefType;
+
+            if (defType != null)
+            {
+                defType.ComputeTypeContainsGCPointers();
+                if (defType.InstanceFieldSize.IsIndeterminate)
+                {
+                    //
+                    // If a method's signature refers to a type with an indeterminate size,
+                    // the compilation will eventually fail when we generate the GCRefMap.
+                    //
+                    // Therefore we need to avoid adding these methods into the graph
+                    //
+                    ThrowHelper.ThrowTypeLoadException(ExceptionStringID.ClassLoadGeneral, type);
+                }
+            }
+
+            ((CompilerTypeSystemContext)type.Context).EnsureLoadableType(type);
         }
 
         private static Instantiation GetInstantiationThatMeetsConstraints(Instantiation definition)
